% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_library.R
\name{all_binary_numbers}
\alias{all_binary_numbers}
\title{Create all possible binary numbers of length 'size', represented in a matrix}
\usage{
all_binary_numbers(size)
}
\arguments{
\item{size}{the length of vector required}
}
\value{
a matrix of all binary numbers of length 'size', with one number
represented as a vector on each row
}
\description{
Create all possible binary numbers of length 'size', represented in a matrix
}
