% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{average_weight_list}
\alias{average_weight_list}
\title{Average weights from multiple neural network runs}
\usage{
average_weight_list(weight_list)
}
\arguments{
\item{weight_list}{A list of neural network weights over multiple iterations. Each main sub-list is a list of neural network weihts
returned from calling get_weights on a keras model}
}
\value{
returns a single weight list, taking the median over all iterations
}
\description{
Average weights from multiple neural network runs
}
