% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_metric_functions.R
\name{calculate_avg_g}
\alias{calculate_avg_g}
\title{Calculates average growth from a tecan file}
\usage{
calculate_avg_g(ods, times, t5_od = 0.46, time_units = "hours",
  by_interval = F)
}
\arguments{
\item{ods}{OD measurements}

\item{times}{time measurements, in seconds}

\item{t5_od}{at what OD has t5 been reached? defaults to 0.46}

\item{time_units}{what to output the average growth by, defaults to hours}

\item{by_interval}{change to True to calculate average growth by interval}
}
\value{
a single number, representing generation time in time_units
}
\description{
Calculates average growth from a tecan file
}
