% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{compare_nn_predictions}
\alias{compare_nn_predictions}
\title{Compare modeled neural network predictions with those observed in XGA with a scatterplot}
\usage{
compare_nn_predictions(nn_model, mapfile, resistance_file, genes_to_predict,
  drug, gene_palette, genes_to_split = c("SNQ2", "PDR5", "YBT1", "YCF1",
  "YOR1"), circle_size = 0.04, text_size = 1.3, text_size_cor = 1.3,
  xlims = c(0, 1), ylims = c(0, 1), axis_labels = T)
}
\arguments{
\item{nn_model}{the neural network model used to make predictions}

\item{mapfile}{a data frame (or matrix) with strain names as row names and genes (optionally, 'Plate'). Same as genotype_file}

\item{resistance_file}{a matrix with strain names as row names and condition names as column}

\item{genes_to_predict}{a vector of genes nn_model was trained with, to use with prediction}

\item{drug}{the drug to compare predictions in (string)}

\item{gene_palette}{a list, with gene names as names, and colours as entries}

\item{genes_to_split}{a vector of genes to average resistance over. Uses the five frequently-associated transporters by default}

\item{circle_size}{arbitrary constant defining point size}

\item{text_size}{constant defining text size in margins, given to cex}

\item{text_size_cor}{constant defining text size in the r value, given to cex}

\item{xlims}{given to plot}

\item{ylims}{given to plot}

\item{axis_labels}{TRUE or FALSE, whether to draw axis labels}
}
\value{
NULL, just makes a plot
}
\description{
Compare modeled neural network predictions with those observed in XGA with a scatterplot
}
