% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_genotype_heatmap.R
\name{create_comparison_matrix}
\alias{create_comparison_matrix}
\title{Creates a matrix comparing resistance of common genotypes of
two separate populations}
\usage{
create_comparison_matrix(genotyping_results_A, genotyping_results_alpha,
  resistance_A, resistance_alpha, genes_of_interest = c("PDR5", "SNQ2",
  "YOR1", "YBT1", "YCF1"))
}
\description{
Creates a matrix comparing resistance of common genotypes of
two separate populations
}
