% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_library.R
\name{create_interaction_matrix}
\alias{create_interaction_matrix}
\title{Return an expanded matrix representing all interaction terms specified}
\usage{
create_interaction_matrix(terms, input_file)
}
\arguments{
\item{terms}{interaction terms and genes as a vector,
split by a ':', e.g. c('A','B','C','A:B','A:C','B:C','A:B:C')}

\item{input_file}{a file which has as the set of all names
defined in the interaction terms as column names
and each row is an observation where each name takes either
0 or 1 as its value}
}
\value{
a matrix which each interaction given is represented
by a separate column. Each interaction term is given by multiplying
the values of its consistuent terms
}
\description{
Return an expanded matrix representing all interaction terms specified
}
