% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_landscape_linear.R
\name{curve_maker_logistic}
\alias{curve_maker_logistic}
\title{Creates a logistic curve between two points}
\usage{
curve_maker_logistic(x1, y1, x2, y2, curve_scale = 0.07, ...)
}
\arguments{
\item{x1, y1, x2, y2}{the beginning and ending co-ordinates}

\item{curve_scale}{defines 'steepness' of the logistic function. Higher values are less steep}
}
\value{
a logistic line on the current graph between the given poins
}
\description{
Creates a logistic curve between two points
}
