% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_landscape_radial.R
\name{draw_radial_landscape}
\alias{draw_radial_landscape}
\title{Smart' adjustment of text size}
\usage{
draw_radial_landscape(nested_list, genes, sample_name, title = "testing",
  color_map = color_map, border_color = rgb(0.1, 0.1, 0.1),
  circle_resolution = 50, title_cex = 3, gene_mode = "knockout",
  text_size_base = 90, max_text_size = 1.5, textcol = NULL,
  max_depth = NULL)
}
\description{
Smart' adjustment of text size
}
