% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_library.R
\name{factor_to_matrix_represenation}
\alias{factor_to_matrix_represenation}
\title{Convert a factor, or a vector of discrete values, to a matrix of zeroes
and ones representing the identity of each level (i.e. a 'one-hot encoding'
matrix)}
\usage{
factor_to_matrix_represenation(terms, input_df)
}
\arguments{
\item{terms}{a vector or factor object, in the same order as it appears
as rows in the 'input_df'}

\item{input_df}{the data frame or matrix from which the vector terms were
extracted from,
with corresponding row names for each measurement}
}
\value{
a 'one-hot' encoding matrix of the factors with the row names supplied
fom input_df
}
\description{
Convert a factor, or a vector of discrete values, to a matrix of zeroes
and ones representing the identity of each level (i.e. a 'one-hot encoding'
matrix)
}
