% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_metrics_library.R
\name{fitness_estimator}
\alias{fitness_estimator}
\title{Estimates a fitness parameter}
\usage{
fitness_estimator(drug, mating_type, frequency_list, time_avail_list,
  metric = "resistance", ctrl = "CTRL", healthy_cutoff = 0.5)
}
\arguments{
\item{drug}{the name of a drug, as it appears in frequency_list and time_avail_list}

\item{mating_type}{the mating type to be measured, as it appears in frequency_list and time_avail_list}

\item{frequency_list}{a list of barcode abundance frequencies of each strain.
The list is first indexed by the drug, then sub-indexed by mating type, then time point
which then gives the vector of all frequencies of a given strain under that drug or time point.
frequency_list must also have a "CTRL" drug which is then indexed directly by mating type rather than time.
this is interpreted as the 'time 0' control}

\item{time_avail_list}{list of available time points, indexed first by drug, then by mating type}

\item{metric}{either 'resistance' or 'growth'.  Resistance returns the difference between growth rate in a given
drug compared to the control (indexed by 'CTRL' in frequency_list), growth returns the estimated growth rate under
a given drug}
}
\value{
a vector of fitness estimates for the measured strains
}
\description{
Estimates a fitness parameter
}
