% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_library.R
\name{get_interactions}
\alias{get_interactions}
\title{Get string corresponding to interactions of degree n amongst given variables}
\usage{
get_interactions(variable_names, degree)
}
\arguments{
\item{variable_names}{the names of variables to be expanded into
interactions}

\item{degree}{the degree of interactions desired}
}
\value{
a vector of strings in A:B:C...etc format
}
\description{
Get string corresponding to interactions of degree n amongst given variables
}
