% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_metric_functions.R
\name{get_time_to_od}
\alias{get_time_to_od}
\title{Obtain the time taken to reach a certain OD in the tecan data}
\usage{
get_time_to_od(od, all_ods, times)
}
\arguments{
\item{od}{the desired OD, for which you want to find out out how long it took to grow to it}

\item{all_ods}{all measured ODs in the data}

\item{times}{timepoint measurements, which should each correspond to all_ods}
}
\value{
a single number, representing an estimate of the time it took to reach the OD
}
\description{
Obtain the time taken to reach a certain OD in the tecan data
}
