% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_metrics_library.R
\name{growth_plotter}
\alias{growth_plotter}
\title{A simple function to plot the growth of a given strain}
\usage{
growth_plotter(drug, mating_type, frequency_list, time_avail_list,
  time_points = NULL, strain_index = 1, col = "black")
}
\arguments{
\item{drug}{the name of a drug, as it appears in frequency_list and time_avail_list}

\item{mating_type}{the mating type to be measured, as it appears in frequency_list and time_avail_list}

\item{frequency_list}{a list of barcode abundance frequencies of each strain.
The list is first indexed by the drug, then sub-indexed by mating type, then time point
which then gives the vector of all frequencies of a given strain under that drug or time point.
frequency_list must also have a "CTRL" drug which is then indexed directly by mating type rather than time.
this is interpreted as the 'time 0' control}

\item{time_avail_list}{list of available time points, indexed first by drug, then by mating type}

\item{strain_index}{the inde(ces) of strain(s) which to make plots for}
}
\description{
A simple function to plot the growth of a given strain
}
