% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_metrics_library.R
\name{make_input_list}
\alias{make_input_list}
\title{Organizes count table into a list for calculating growth metrics}
\usage{
make_input_list(input_file, drugs = NULL, mating_types = c("A", "alpha"),
  tags = c("UP", "DN"), time_points = c("5", "10", "15", "20"))
}
\arguments{
\item{input_file}{count table}

\item{drugs}{drugs in input_file, inferred automatically if NULL}

\item{mating_types}{mating types to be processed (A/alpha)}

\item{tags}{tags to be processed (UP/DN)}

\item{time_points}{timepoints to be processed (5,10,15,20)}
}
\value{
a list with time points, t0, and counts for each drug, mating type, and tag combination
}
\description{
Organizes count table into a list for calculating growth metrics
}
