% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{make_nn_model}
\alias{make_nn_model}
\title{named_weights' = model_weights,}
\usage{
make_nn_model(resistance_file, genotype_file, condition_names, genes = NULL,
  effect_size_threshold = 0, regularization_rate = 1e-04,
  learning_rate = 0.01, epochs = 2000, batch_size = 2000,
  validation_split = 0.1, act_type = "sigmoid", train_model = T,
  efflux_genes = NULL)
}
\description{
named_weights' = model_weights,
}
