% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_landscape.R
\name{mata_matalpha_comparison_histogram}
\alias{mata_matalpha_comparison_histogram}
\title{Plot a histogram of grouped resistance measurement correlations
between MATa and MATalpha pools}
\usage{
mata_matalpha_comparison_histogram(A_resistance_file = NULL,
  alpha_resistance_file = NULL, A_genotyping_df = NULL,
  alpha_genotyping_df = NULL, genes = c("PDR5", "SNQ2", "YOR1", "YBT1",
  "YCF1"), xlims = NULL, breaks = "Sturges")
}
\arguments{
\item{A_resistance_file}{}

\item{alpha_resistance_file}{}

\item{A_genotyping_df}{genotyping data for MAT}

\item{alpha_genotyping_df}{}

\item{genes}{}

\item{xlims}{}

\item{breaks}{}
}
\description{
Plot a histogram of grouped resistance measurement correlations
between MATa and MATalpha pools
}
