% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_landscape.R
\name{metric_comparison_scatterplot}
\alias{metric_comparison_scatterplot}
\title{Title}
\usage{
metric_comparison_scatterplot(od_list, concentration = NULL, split_df,
  A_resistance_file = NULL, alpha_resistance_file = NULL,
  A_genotyping_df = NULL, alpha_genotyping_df = NULL,
  mating_types = c("A", "alpha"), drug = "fluconazole", metric = "IC50",
  gene_palette = list(PDR5 = rgb(47, 144, 206, maxColorValue = 255), SNQ2 =
  rgb(179, 231, 172, maxColorValue = 255), YOR1 = rgb(233, 153, 76,
  maxColorValue = 255), YBT1 = rgb(166, 7, 13, maxColorValue = 255), YCF1 =
  rgb(255, 255, 191, maxColorValue = 255)), circle_border_color = "black",
  xlims = NULL, ylims = NULL, circle_scale_x = 0.04,
  circle_scale_y = 0.04, legend = T)
}
\arguments{
\item{legend}{}
}
\description{
Title
}
