% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{nn_model_search}
\alias{nn_model_search}
\title{Searches over regularization rates with the neural network training procedure}
\usage{
nn_model_search(resfile, mapfile, condition_names, genes,
  regularization_rates = 2e-04, learning_rate = 0.05, epochs = 10000,
  batch_size = 1000, runs_averaged = 10, n_clusters = 10,
  validation_split = 0.1, nn_function = make_nn_model,
  weight_merging_function = prune_weight_list, efflux_genes = NULL,
  regularization_rates_indirect = NULL, three_layer_model = F,
  searching_indirect = F)
}
\arguments{
\item{resfile}{a matrix with strain names as row names and condition names as column (same as resistance_file)}

\item{mapfile}{a data frame (or matrix) with strain names as row names and genes (optionally, 'Plate'). Same as genotype_file}

\item{condition_names}{what to name conditions in the neural network (vector of strings)}

\item{genes}{what genes the neural network models in the first layer (and second layer by default)
defaults to column names 2-17 of genotype_file (corresponding to names of all 16 transporters)}

\item{regularization_rates}{what L1 regularzation rates to search for}

\item{learning_rate}{rate for regularizer_l1 passed to keras (L1 regularization rate)}

\item{epochs}{epochs passed to keras; defines number of training epochs}

\item{batch_size}{batch_size passed to keras; defines how many examples to sample in gradient descent}

\item{runs_averaged}{number of runs to combine}

\item{n_clusters}{how many neural networks to train in parallel}

\item{validation_split}{validation_split passed to keras}

\item{nn_function}{function used to train a single neural network (defaults to make_nn_model)}

\item{weight_merging_function}{function used to merge weights defaults to prune_weight_list}

\item{efflux_genes}{vector, what genes the neural network models in the second layer}

\item{regularization_rates_indirect}{vector, what 'indirect' L1 regularization rates to search over (if null, just sets the indirect rates the same as the direct_rates)}

\item{three_layer_model}{TRUE or FALSE, is this a three_layer_model? (i.e. do we use the regulation_rates_indirect argument)}

\item{searching_indirect}{TRUE or FALSE,  do we search over regularization_rates_indirect? (if FALSE, just give one argument to regularization_rates_indirect}
}
\value{
a data frame (or list of data frames if searching indirect) with the following column names:
'reg_rate', the L1 regularization rate used
'mse_initial' the mean-squared error of the neural network before significance testing of weights
'cor_initial' correlation between predicted and measured resistances before significance testing of weights
'n_param_initial' the number of non-zero weights before significance testing
'sum_weights_initial' the sum of absoulte neural network weights before significance testing
'mse_final' the mean-squared error of the neural network after significance testing of weights
'cor_final' correlation between predicted and measured resistances after significance testing of weights
'n_param_final' the number of non-zero weights after significance testing
'sum_weights_final' the sum of absoulte neural network weights after significance testing
}
\description{
Searches over regularization rates with the neural network training procedure
}
