% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_metric_functions.R
\name{parse_ods_tecan}
\alias{parse_ods_tecan}
\title{Parse tecan file into an R-readable form}
\usage{
parse_ods_tecan(tecan_file)
}
\arguments{
\item{tecan_file}{a .txt output of a tecan file}
}
\value{
a dataframe with all time measurements and ODs
}
\description{
Parse tecan file into an R-readable form
}
