% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_metrics_library.R,
%   R/tecan_metric_functions.R
\name{rhombus_integration}
\alias{rhombus_integration}
\alias{rhombus_integration}
\alias{rhombus_integration}
\title{Perform two-dimensional numerical integration using the rhombus approximation}
\usage{
rhombus_integration(x, y)

rhombus_integration(x, y)

rhombus_integration(x, y)
}
\arguments{
\item{x}{vector of x values ('lengths')}

\item{y}{vector y values ('heights')}

\item{x}{vector of x values ('lengths')}

\item{y}{vector y values ('heights')}
}
\value{
area under the polygon formed by x,y

area under the polygon formed by x,y
}
\description{
Perform two-dimensional numerical integration using the rhombus approximation

Perform two-dimensional numerical integration using the rhombus approximation

A generic integration function you can use to calculate AUC
}
