% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_library.R
\name{set_overlap_p}
\alias{set_overlap_p}
\title{Calculates p value of an overlap between two sets}
\usage{
set_overlap_p(n_overlap, set_size1, set_size2, universe_size)
}
\arguments{
\item{n_overlap}{how many objects the two sets have in common}

\item{set_size1}{how many objects in the first set}

\item{set_size2}{how many objects in the second set}

\item{universe_size}{how many objects in the 'universe', i.e. the size
of all possible things both sets could have sampled}
}
\value{
a p value calculated using the Hypergeometric test
}
\description{
Calculates p value of an overlap between two sets
}
