% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_landscape_linear.R
\name{split_df_to_list}
\alias{split_df_to_list}
\title{Split data frame to list}
\usage{
split_df_to_list(binary_df, binary_colnames, base_name = "wt")
}
\arguments{
\item{binary_df}{a data frame containing at least one column which is binary variables (i.e. values are either 0 or 1)}

\item{binary_colnames}{names of columns containing binary variables between which to split}
}
\value{
returns a list where each entry is a subset of binary_df where a given
combination of binary variables are all 1. The names represent the binary variables
which are set to equal 1 in each subset
}
\description{
Split data frame to list
}
