% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_library.R
\name{stepwise_feature_elimination_glm}
\alias{stepwise_feature_elimination_glm}
\title{Perform stepwise feature elimination on a general linear regression model using
Type III SS Anova}
\usage{
stepwise_feature_elimination_glm(my_glm, alpha = 0.05, given_data = NULL)
}
\arguments{
\item{my_glm}{a general linear model object}

\item{alpha}{the minimum p value of variables to keep}

\item{given_data}{the source data on which my_glm was trained on, if NULL (default)
tries to extract it from the lm object}
}
\value{
a new general linear model with all variables with significance less than alpha
 removed in a stepwise fashion
}
\description{
Perform stepwise feature elimination on a general linear regression model using
Type III SS Anova
}
