% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_library.R
\name{stepwise_feature_elimination_lm}
\alias{stepwise_feature_elimination_lm}
\title{Perform stepwise feature elimination on a linear regression model using
Type III SS Anova}
\usage{
stepwise_feature_elimination_lm(my_lm, alpha = 0.05, given_data = NULL)
}
\arguments{
\item{my_lm}{a linear model object}

\item{alpha}{the minimum p value of variables to keep}

\item{given_data}{the source data on which my_lm was trained on, if NULL
(default) tries to extract it from the lm object}
}
\value{
a new linear model with all variables with significance less than
alpha removed in a stepwise fashion
}
\description{
Perform stepwise feature elimination on a linear regression model using
Type III SS Anova
}
